Fix sqlsrv for LogoutStore version migrations
Fixes https://github.com/simplesamlphp/simplesamlphp/issues/1379

This fixes the incorrect `ALTER ... MODIFY` syntax for SQLServer (it's `ALTER ... ALTER`) which prevents the LogoutStore to migration from version `1 -> 2` and from version `3 -> 4`. Previously, migrating the version failed and the old version was kept. This caused a warning to be logged but no other effect from what I could see.

I used https://github.com/simplesamlphp/docker-simplesamlphp along with a SQLServer image to verify that this fixes the problem, and I was able to successfully migrate the LogoutStore from version 1 -> 2 and from version 3 -> 4 (2 -> 3 worked before this fix).